refactor: extract agent identity helpers
